% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl.R
\name{BlSpline}
\alias{BlSpline}
\title{P-Spline base learner class}
\description{
This class creates a base learner that can be used for calculating
predictions, fitting parameters, etc.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BlSpline$new()}}
\item \href{#method-initData}{\code{BlSpline$initData()}}
\item \href{#method-initDataXtX}{\code{BlSpline$initDataXtX()}}
\item \href{#method-train}{\code{BlSpline$train()}}
\item \href{#method-predict}{\code{BlSpline$predict()}}
\item \href{#method-predictNewdata}{\code{BlSpline$predictNewdata()}}
\item \href{#method-predictVal}{\code{BlSpline$predictVal()}}
\item \href{#method-updateParam}{\code{BlSpline$updateParam()}}
\item \href{#method-getKnotRange}{\code{BlSpline$getKnotRange()}}
\item \href{#method-getFeature}{\code{BlSpline$getFeature()}}
\item \href{#method-getHyperpars}{\code{BlSpline$getHyperpars()}}
\item \href{#method-getParam}{\code{BlSpline$getParam()}}
\item \href{#method-clone}{\code{BlSpline$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlSpline$new(knots_min, knots_max, nknots = 20L, ord = 3L, derivs = 2L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{knots_min}}{(`numeric(1L)`)\cr
Minimum number of inner knots.}

\item{\code{knots_max}}{(`numeric(1L)`)\cr
Maximum number of inner knots.}

\item{\code{nknots}}{(`integer(1L)`)\cr
Number of inner knots.}

\item{\code{ord}}{(`integer(1L)`)\cr
Polynomial degree of the basis functions.}

\item{\code{derivs}}{(`integer(1L)`)\cr
Number of penalized differences.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initData"></a>}}
\if{latex}{\out{\hypertarget{method-initData}{}}}
\subsection{Method \code{initData()}}{
Initializes data based on given data. This function should
be used at the sites.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlSpline$initData(x, feature, df, val_idx = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(`numeric()`)\cr
Data vector.}

\item{\code{feature}}{(`character(1L)`)\cr
Character containing the name of the feature.}

\item{\code{df}}{(`numeric(1L)`)\cr
Value indicating the degrees of freedom for the base learner.}

\item{\code{val_idx}}{(`integer()`)\cr
Vector containing the indices used for validation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initDataXtX"></a>}}
\if{latex}{\out{\hypertarget{method-initDataXtX}{}}}
\subsection{Method \code{initDataXtX()}}{
Initializes data based on the hat matrix. This function should
be used at the host.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlSpline$initDataXtX(feature, df, xtx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{(`character(1L)`)\cr
Character containing the name of the feature.}

\item{\code{df}}{(`numeric(1L)`)\cr
Value indicating the degrees of freedom for the base learner.}

\item{\code{xtx}}{(`matrix()`)\cr
Hat matrix used for parameter estimation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\if{latex}{\out{\hypertarget{method-train}{}}}
\subsection{Method \code{train()}}{
Calculate parameter estimates based on
be used at the host.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlSpline$train(y, xty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{(`numeric(1L)`)\cr
Value indicating the degrees of freedom for the base learner.}

\item{\code{xty}}{(`matrix()|numeric()`)\cr
Hat matrix used for parameter estimation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Calculate prediction based on a parameter vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlSpline$predict(par = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{par}}{(`numeric()`)\cr
Parameter vector used for predicting on the data set
in design. If `NULL`, the internal parameter vector is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predictNewdata"></a>}}
\if{latex}{\out{\hypertarget{method-predictNewdata}{}}}
\subsection{Method \code{predictNewdata()}}{
Calculate prediction for the new data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlSpline$predictNewdata(newdata, par = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{(`data.frame()`)\cr
New data. Note that `newdata` must contain the feature.}

\item{\code{par}}{(`numeric()`)\cr
Parameter vector used for predicting on new data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predictVal"></a>}}
\if{latex}{\out{\hypertarget{method-predictVal}{}}}
\subsection{Method \code{predictVal()}}{
Calculate prediction for the validation data based on a
parameter vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlSpline$predictVal(par = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{par}}{(`numeric()`)\cr
Parameter vector used for predicting on the data set
in validation design. If `NULL`, the internal parameter
vector is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateParam"></a>}}
\if{latex}{\out{\hypertarget{method-updateParam}{}}}
\subsection{Method \code{updateParam()}}{
Add parameter vector to the already stored parameter vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlSpline$updateParam(par)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{par}}{(`numeric()`)\cr
Parameter vector added to the internal stored parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getKnotRange"></a>}}
\if{latex}{\out{\hypertarget{method-getKnotRange}{}}}
\subsection{Method \code{getKnotRange()}}{
Get knot range.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlSpline$getKnotRange()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFeature"></a>}}
\if{latex}{\out{\hypertarget{method-getFeature}{}}}
\subsection{Method \code{getFeature()}}{
Get feature name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlSpline$getFeature()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHyperpars"></a>}}
\if{latex}{\out{\hypertarget{method-getHyperpars}{}}}
\subsection{Method \code{getHyperpars()}}{
Get hyperparameters (df, penalty, order, and derivs).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlSpline$getHyperpars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getParam"></a>}}
\if{latex}{\out{\hypertarget{method-getParam}{}}}
\subsection{Method \code{getParam()}}{
Get parameter vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlSpline$getParam()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlSpline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
