% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{checkConnection}
\alias{checkConnection}
\title{Check the DataSHIELD connection}
\usage{
checkConnection(connections)
}
\arguments{
\item{connections}{(`list(OpalConnection)`)\cr
Connections to the DataSHIELD servers.}
}
\description{
Check the DataSHIELD connection
}
