% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{checkSymbol}
\alias{checkSymbol}
\title{Check if symbol is present in global envir.}
\usage{
checkSymbol(symbol)
}
\arguments{
\item{symbol}{(`character(1L)`)\cr
Symbol specifying something.}
}
\description{
Check if symbol is present in global envir.
}
