% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-funs.R
\name{createClientModel}
\alias{createClientModel}
\title{Initialize a client model}
\usage{
createClientModel(
  symbol,
  target,
  feature_names = NULL,
  learning_rate = 0.1,
  df = 5,
  nknots = 20L,
  ord = 3L,
  derivs = 2L,
  val_fraction = NULL,
  positive = NULL,
  seed = NULL
)
}
\arguments{
\item{symbol}{(`character(1L)`)\cr
Character containing the name of the data.}

\item{target}{(`character(1L)`)\cr
Character containing the name of the target variable.}

\item{feature_names}{(`character()`)\cr
Character vector of all target variables.}

\item{learning_rate}{(`numeric(1L)`)\cr
Learning rate.}

\item{df}{(`numeric(1L)`)\cr
Degrees of freedom.}

\item{nknots}{(`integer(1L)`)\cr
Number of inner knots.}

\item{ord}{(`integer(1L)`)\cr
Polynomial degree of basis functions.}

\item{derivs}{(`integer(1L)`)\cr
Number of penalized differences.}

\item{val_fraction}{(`numeric(1L)`)\cr
Fraction of samples used for validation.}

\item{positive}{(`character(1L)`)\cr
Name of the positive class in binary classification.}

\item{seed}{(`numeric(1L)`)\cr
Seed for generating validation data (only applies when val_fraction is set).}
}
\value{
Client model of R6 class ClientModel.
}
\description{
Initialize a client model
}
