% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{decodeBinary}
\alias{decodeBinary}
\title{Deserialize object}
\usage{
decodeBinary(bin, package = NULL)
}
\arguments{
\item{bin}{(`character(1L)`) Binary string value containing the serialized model.}

\item{package}{(`character(1L)`) Package required for object deserialization (default is `NULL`).}
}
\value{
Deserialized object from `bin`
}
\description{
Decode a given string of a serialized object.
}
\examples{
mod = lm(Sepal.Width ~ ., data = iris)
bin = encodeObject(mod)
mod_b = decodeBinary(bin)
all.equal(mod, mod_b)
}
\author{
Daniel S.
}
