% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{deparseParams}
\alias{deparseParams}
\title{Deparse a character vector representation}
\usage{
deparseParams(xchar, sep = "/")
}
\arguments{
\item{xchar}{(`character(1L)`)\cr
Character string obtained by `parseParams`.}

\item{sep}{(`character(1L)`)\cr
Character used to separate numerical values.}
}
\description{
Converts a string created by `parseParams` into the corresponding
  numeric vector.
}
\examples{
x = rnorm(20)
xchar = parseParams(x)
xnew = deparseParams(xchar)
all.equal(x, xnew)
}
