% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{dsNaRm}
\alias{dsNaRm}
\title{Remove all rows that contain missing values.}
\usage{
dsNaRm(symbol)
}
\arguments{
\item{symbol}{(`character(1L)`)\cr
Symbol specifying something.}
}
\description{
Remove all rows that contain missing values.
}
