% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-funs.R
\name{getBlHyperpars}
\alias{getBlHyperpars}
\title{Update the penalties of the client models.}
\usage{
getBlHyperpars(model_symbol)
}
\arguments{
\item{model_symbol}{(`character(1L)`)\cr
The name of the model at the server.}
}
\description{
Update the penalties of the client models.
}
