% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-funs.R
\name{getClientModelCoefficients}
\alias{getClientModelCoefficients}
\title{Get the model coefficients.}
\usage{
getClientModelCoefficients(model_symbol = "cwb")
}
\arguments{
\item{model_symbol}{(`character(1L)`)\cr
The name of the model at the server.}
}
\value{
Model coefficients per base learner.
}
\description{
Get the model coefficients.
}
