% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getClientNrow}
\alias{getClientNrow}
\title{Get number of rows of client data}
\usage{
getClientNrow(symbol)
}
\arguments{
\item{symbol}{(`character(1L)`)\cr
Symbol specifying something.}
}
\description{
Get number of rows of client data
}
