% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getClientTaskType}
\alias{getClientTaskType}
\title{Get the task type.}
\usage{
getClientTaskType(symbol, target)
}
\arguments{
\item{symbol}{(`character(1L)`)\cr
Symbol specifying something.}

\item{target}{(`character(1L)`)\cr
The name response variable.}
}
\description{
Get the task type.
}
