% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-funs.R
\name{getClientTrainValObs}
\alias{getClientTrainValObs}
\title{Get the number of rows per train and validation set.}
\usage{
getClientTrainValObs(model_symbol = "cwb")
}
\arguments{
\item{model_symbol}{(`character(1L)`)\cr
The name of the model at the server.}
}
\value{
Number of observation in the train and validation set.
}
\description{
Get the number of rows per train and validation set.
}
