% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-funs.R
\name{getClientXtX}
\alias{getClientXtX}
\title{Get XtX from the client models.}
\usage{
getClientXtX(model_symbol = "cwb")
}
\arguments{
\item{model_symbol}{(`character(1L)`)\cr
The name of the model at the server.}
}
\value{
XtX matrices of the registered base learners.
}
\description{
Get XtX from the client models.
}
