% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-funs.R
\name{getClientXty}
\alias{getClientXty}
\title{Get Xty from the client models.}
\usage{
getClientXty(model_symbol = "cwb")
}
\arguments{
\item{model_symbol}{(`character(1L)`)\cr
The name of the model at the server.}
}
\value{
Xty matrices of the registered base learners.
}
\description{
Get Xty from the client models.
}
