% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getDataSHIELDInfo}
\alias{getDataSHIELDInfo}
\title{Get the session information of the DataSHIELD server}
\usage{
getDataSHIELDInfo()
}
\value{
list of session infos returned from `sessionInfo()` of each machine
}
\description{
This method returns `sessionInfo()` from the used DataSHIELD servers.
  The main purpose is for testing and checking the environment used on the remote servers.
}
\author{
Daniel S.
}
