% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-funs.R
\name{getFeatureNames}
\alias{getFeatureNames}
\title{Get feature names from sites.}
\usage{
getFeatureNames(model_symbol)
}
\arguments{
\item{model_symbol}{(`character(1L)`)\cr
The name of the model at the server.}
}
\description{
Get feature names from sites.
}
