% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-funs.R
\name{getOptimalConstant}
\alias{getOptimalConstant}
\title{Get loss optimal constants per site.}
\usage{
getOptimalConstant(model_symbol)
}
\arguments{
\item{model_symbol}{(`character(1L)`)\cr
The name of the model at the server.}
}
\description{
Get loss optimal constants per site.
}
