% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-funs.R
\name{getRisk}
\alias{getRisk}
\title{Get risk from model.}
\usage{
getRisk(model_symbol, risk_type = "train")
}
\arguments{
\item{model_symbol}{(`character(1L)`)\cr
The name of the model at the server.}

\item{risk_type}{(`character(1L)`)\cr
The risk type, must be one of "full", "train" (default), or "val".}
}
\description{
Get risk from model.
}
