% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{getTaskType}
\alias{getTaskType}
\title{Get the task type of a vector}
\usage{
getTaskType(x)
}
\arguments{
\item{x}{(`vector()`)\cr
Vector from which the task is extracted. Just regression and binary classification is supported.}
}
\value{
Character containing the task type.
}
\description{
Get the task type of a vector
}
