% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-funs.R
\name{initClientModel}
\alias{initClientModel}
\title{Initialize the site base learner.}
\usage{
initClientModel(model_symbol, ll_init_binary)
}
\arguments{
\item{model_symbol}{(`character(1L)`)\cr
The name of the model at the server.}

\item{ll_init_binary}{(`character(1L)`)\cr
Binary encoded list of initialization values.}
}
\description{
Initialize the site base learner.
}
