% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-funs.R
\name{initSiteConstant}
\alias{initSiteConstant}
\title{Initialize client model with constant.}
\usage{
initSiteConstant(model_symbol, init)
}
\arguments{
\item{model_symbol}{(`character(1L)`)\cr
The name of the model at the server.}

\item{init}{(`numeric(1L)`)\cr
Constant initialization.}
}
\description{
Initialize client model with constant.
}
