% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{LossBinomial}
\alias{LossBinomial}
\title{Binomial loss class}
\description{
This class defines the loss and the specific functions required for boosting.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-init}{\code{LossBinomial$init()}}
\item \href{#method-loss}{\code{LossBinomial$loss()}}
\item \href{#method-pseudoResids}{\code{LossBinomial$pseudoResids()}}
\item \href{#method-risk}{\code{LossBinomial$risk()}}
\item \href{#method-aggregateInit}{\code{LossBinomial$aggregateInit()}}
\item \href{#method-clone}{\code{LossBinomial$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
Get the constant initialization of the loss.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossBinomial$init(truth)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{truth}}{(`numeric()`)\cr
Response vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loss"></a>}}
\if{latex}{\out{\hypertarget{method-loss}{}}}
\subsection{Method \code{loss()}}{
Calculate the pointwise loss.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossBinomial$loss(truth, pred)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{truth}}{(`numeric()`)\cr
The response variable.}

\item{\code{pred}}{(`numeric()`)\cr
The predicted values for the response.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pseudoResids"></a>}}
\if{latex}{\out{\hypertarget{method-pseudoResids}{}}}
\subsection{Method \code{pseudoResids()}}{
Calculate the pseudo residuals.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossBinomial$pseudoResids(truth, pred)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{truth}}{(`numeric()`)\cr
The response variable.}

\item{\code{pred}}{(`numeric()`)\cr
The predicted values for the response.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-risk"></a>}}
\if{latex}{\out{\hypertarget{method-risk}{}}}
\subsection{Method \code{risk()}}{
Calculate the risk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossBinomial$risk(truth, pred)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{truth}}{(`numeric()`)\cr
The response variable.}

\item{\code{pred}}{(`numeric()`)\cr
The predicted values for the response.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aggregateInit"></a>}}
\if{latex}{\out{\hypertarget{method-aggregateInit}{}}}
\subsection{Method \code{aggregateInit()}}{
Aggregate initializations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossBinomial$aggregateInit(inits, w)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inits}}{(`numeric()`)\cr
Initialization values.}

\item{\code{w}}{(`numeric()`)\cr
Weights.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossBinomial$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
