% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{parseParams}
\alias{parseParams}
\title{Paste a numeric vector into a character vector}
\usage{
parseParams(x, sep = "/", digits = 22L)
}
\arguments{
\item{x}{(`numeric()`)\cr
Numeric vector converted into the string.}

\item{sep}{(`character(1L)`)\cr
Character used to separate numerical values.}

\item{digits}{(`integer(1L)`)\cr
Significant digits used in the string.}
}
\description{
Converts a numerical vector into one string with the values separated
  by the character specified in `sep` (default = /).
}
\examples{
x = rnorm(20)
xchar = parseParams(x)
}
