% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bltraces.R
\name{plotBaselearnerTraces}
\alias{plotBaselearnerTraces}
\title{Visualize base learner traces}
\usage{
plotBaselearnerTraces(cwb, value = 1, n_legend = 5L, add_effect_type = TRUE)
}
\arguments{
\item{value}{(`numeric(1L) | numeric()`)\cr
Value used to show the base learner development w.r.t. to the value.}

\item{n_legend}{(`integer(1L)`)\cr
Number of colored base learners added to the legend.}

\item{cboost}{(`HostModel`)\cr
A trained host model.}
}
\value{
\code{ggplot} object containing the graphic.
}
\description{
This function shows how the base learners evolves over the fitting process.
The default is to show how the frequency of a single base learner included
into the model evolves. Additionally, with the \code{value} argument, vectors
(e.g. the risk) can be used to show how the base learner specific risk reduction
evolves during the fitting process.
}
