% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe-data.R
\name{sharedFEDataCat}
\alias{sharedFEDataCat}
\title{Get data for a categorical shared effect}
\usage{
sharedFEDataCat(cwb, feature)
}
\arguments{
\item{cwb}{()\cr}

\item{feature}{(`character(1L)`)\cr
Name of the feature.}
}
\value{
`data.frame()` with feature values and the prediction.
}
\description{
Get data for a categorical shared effect
}
