% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe-data.R
\name{sharedFEDataNum}
\alias{sharedFEDataNum}
\title{Get data for a numerical shared effect}
\usage{
sharedFEDataNum(cwb, feature, npoints = 100L)
}
\arguments{
\item{cwb}{()\cr}

\item{feature}{(`character(1L)`)\cr
Name of the feature.}

\item{npoints}{(`integer(1L)`)\cr
Number of feature points (equidistand grid) used for predicting.}
}
\value{
`data.frame()` with feature values and the prediction.
}
\description{
Get data for a numerical shared effect
}
