% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe-data.R
\name{siteFEDataCat}
\alias{siteFEDataCat}
\title{Get data for a categorical site effect}
\usage{
siteFEDataCat(cwb, feature)
}
\arguments{
\item{cwb}{()\cr}

\item{feature}{(`character(1L)`)\cr
Name of the feature.}
}
\value{
`data.frame()` with feature values, the prediction, and the server.
}
\description{
Get data for a categorical site effect
}
