% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe-data.R
\name{siteFEDataNum}
\alias{siteFEDataNum}
\title{Get data for a numerical site effect}
\usage{
siteFEDataNum(cwb, feature, npoints = 100L)
}
\arguments{
\item{cwb}{()\cr}

\item{feature}{(`character(1L)`)\cr
Name of the feature.}

\item{npoints}{(`integer(1L)`)\cr
Number of feature points (equidistand grid) used for predicting.}
}
\value{
`data.frame()` with feature values, the prediction, and the server.
}
\description{
Get data for a numerical site effect
}
