% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl-trafos.R
\name{trafoCategorical}
\alias{trafoCategorical}
\title{Create dummy data matrix}
\usage{
trafoCategorical(x, levels)
}
\arguments{
\item{x}{(`character()`) Data vector.}

\item{levels}{(`character()`) Levels of the character vector.}
}
\value{
Sparse dummy matrix..
}
\description{
Create dummy data matrix
}
\author{
Daniel S.
}
