% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl-trafos.R
\name{trafoSpline}
\alias{trafoSpline}
\title{Create a B-spline basis representation}
\usage{
trafoSpline(x, knots_min, knots_max, nknots = 20L, ord = 3L)
}
\arguments{
\item{x}{(`numeric()`) Data vector.}

\item{knots_min}{(`numeric(1L`) Minimal value for the knots.}

\item{knots_max}{(`numeric(1L`) Maximal value for the knots.}

\item{nknots}{(`integer(1L)`) Number of knots.}

\item{ord}{(`integer(1L)`) Basis function degree.}
}
\value{
Sparse matrix containing the spline basis.
}
\description{
Create a B-spline basis representation
}
\author{
Daniel S.
}
