% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-funs.R
\name{updateClientBaselearner}
\alias{updateClientBaselearner}
\title{Update model base learner}
\usage{
updateClientBaselearner(model_symbol, blname, par_binary = NULL)
}
\arguments{
\item{model_symbol}{(`character(1L)`)\cr
The name of the model at the server.}

\item{blname}{(`character(1L)`)\cr
Name of the base learner.}

\item{par_binary}{(`character(1L)`)\cr
Binary encoded parameter used to update pseudo residuals and predictions.}
}
\description{
Update model base learner
}
