% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-funs.R
\name{updateClientPenalty}
\alias{updateClientPenalty}
\title{Update the penalties of the client models.}
\usage{
updateClientPenalty(model_symbol, ll_pen_binary, anistrop = TRUE)
}
\arguments{
\item{model_symbol}{(`character(1L)`)\cr
The name of the model at the server.}

\item{ll_pen_binary}{(`character(1L)`)\cr
Named list with new penalty terms encoded by `encodeObject`.}

\item{anistrop}{(`logical(1L)`)\cr
Flag indicating whether the penalty should be done anistrop or isotrop.}
}
\description{
Update the penalties of the client models.
}
