% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-funs.R
\name{getBlHyperpars}
\alias{getBlHyperpars}
\title{Get the hyperaprameter of the client base learners.}
\usage{
getBlHyperpars(model_symbol)
}
\arguments{
\item{model_symbol}{(`character(1L)`)\cr
The name of the model at the server.}
}
\description{
Get the hyperaprameter of the client base learners.
}
