% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-funs.R
\name{getClientInit}
\alias{getClientInit}
\title{Get base learner initialization}
\usage{
getClientInit(symbol, feature_names)
}
\arguments{
\item{symbol}{(`character(1L)`)\cr
Character containing the name of the data.}

\item{feature_names}{(`character()`)\cr
Character vector of all target variables.}
}
\value{
List with initialization values.
}
\description{
Get base learner initialization
}
