% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-model.R
\name{HostModel}
\alias{HostModel}
\title{Host model}
\description{
This class creates a CWB model container for the host.
It collects all base learners.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{loss}}{(`Loss`)\cr
Loss object used for modelling.}

\item{\code{lr}}{(numeric(1L)\cr
Learning rate.}

\item{\code{bls}}{(`list(Baselearner)`)\cr
List of base learner.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HostModel$new()}}
\item \href{#method-addBaselearners}{\code{HostModel$addBaselearners()}}
\item \href{#method-getDataSymbol}{\code{HostModel$getDataSymbol()}}
\item \href{#method-getFeatureNames}{\code{HostModel$getFeatureNames()}}
\item \href{#method-getParam}{\code{HostModel$getParam()}}
\item \href{#method-update}{\code{HostModel$update()}}
\item \href{#method-log}{\code{HostModel$log()}}
\item \href{#method-clone}{\code{HostModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HostModel$new(
  symbol = "D",
  target,
  target_type = NULL,
  feature_names = NULL,
  learning_rate = 0.1,
  df = 5L,
  nknots = 20L,
  ord = 3L,
  derivs = 2L,
  positive = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{(`character(1L)`)\cr
Character containing the name of the data.}

\item{\code{target}}{(`character(1L)`)\cr
Character containing the name of the target variable.}

\item{\code{target_type}}{(`character(1L)`)\cr
The target type, must be `regression` or `bin-class`.}

\item{\code{feature_names}}{(`character()`)\cr
Character vector of all target variables.}

\item{\code{learning_rate}}{(`numeric(1L)`)\cr
Learning rate.}

\item{\code{df}}{(`numeric(1L)`)\cr
Degrees of freedom.}

\item{\code{nknots}}{(`integer(1L)`)\cr
Number of inner knots.}

\item{\code{ord}}{(`integer(1L)`)\cr
Polynomial degree of basis functions.}

\item{\code{derivs}}{(`integer(1L)`)\cr
Number of penalized differences.}

\item{\code{positive}}{(`integer(1L)`)\cr
Character indicating the positive class in the binary classification setting.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addBaselearners"></a>}}
\if{latex}{\out{\hypertarget{method-addBaselearners}{}}}
\subsection{Method \code{addBaselearners()}}{
Add base learners to the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HostModel$addBaselearners(ll_init, ll_xtx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ll_init}}{(`list()`)\cr
List containing the initialization parameters.}

\item{\code{ll_xtx}}{(`list()`)\cr
List containing the XtX matrices.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDataSymbol"></a>}}
\if{latex}{\out{\hypertarget{method-getDataSymbol}{}}}
\subsection{Method \code{getDataSymbol()}}{
Get the data symbol
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HostModel$getDataSymbol()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFeatureNames"></a>}}
\if{latex}{\out{\hypertarget{method-getFeatureNames}{}}}
\subsection{Method \code{getFeatureNames()}}{
Get feature names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HostModel$getFeatureNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getParam"></a>}}
\if{latex}{\out{\hypertarget{method-getParam}{}}}
\subsection{Method \code{getParam()}}{
Get parameter based on xty.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HostModel$getParam(ll_xty)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ll_xty}}{(`list()`)\cr
List containing the Xty matrices.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Conduct an update step.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HostModel$update(blname, par)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blname}}{(`character(1L)`)\cr
Base learner name.}

\item{\code{par}}{(`numeric()`)\cr
Parameter vector corresponding to blname.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log"></a>}}
\if{latex}{\out{\hypertarget{method-log}{}}}
\subsection{Method \code{log()}}{
Log an updated step.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HostModel$log(blname, effect_type, sse, risk_train = NA, risk_val = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blname}}{(`character(1L)`)\cr
Base learner name.}

\item{\code{effect_type}}{(`character(1L)`)\cr
Type (shared `shared` or site-specific `site`) of the base learner.}

\item{\code{sse}}{(`numeric(1L)`)\cr
SSE of the added base learner.}

\item{\code{risk_train}}{(`numeric(1L)`)\cr
Training risk.}

\item{\code{risk_val}}{(`numeric(1L)`)\cr
Validation risk.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HostModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
